CSP support for `DockerManagementServer` (#1107)
* CSP support for `DockerManagementServer`

* Spotless
