+++ src/main/java/com/nirima/jenkins/plugins/docker/utils/Consts.java
@@ -10,9 +10,4 @@ public final class Consts {
      * The base URL of the plugin images.
      */
     public static final String PLUGIN_IMAGES_URL = PLUGIN_URL + "images/";
-
-    /**
-     * The base URL of the plugin javascripts.
-     */
-    public static final String PLUGIN_JS_URL = PLUGIN_URL + "js/";
 }
