+++ src/main/resources/com/nirima/jenkins/plugins/docker/DockerManagementServer/index.jelly
@@ -1,9 +1,9 @@
 <?jelly escape-by-default='true'?>
-<j:jelly xmlns:j="jelly:core" xmlns:l="/lib/layout"  xmlns:f="/lib/form"
+<j:jelly xmlns:j="jelly:core" xmlns:st="jelly:stapler" xmlns:l="/lib/layout"  xmlns:f="/lib/form"
         >
     <l:layout title="${%Docker hosts}" norefresh="true" permission="${it.requiredPermission}">
         <l:header>
-            <script src="${rootURL}${it.getJsUrl('docker-manage.js')}" type="text/javascript"></script>
+            <st:adjunct includes="com.nirima.jenkins.plugins.docker.DockerManagementServer.docker-manage"/>
             <style type="text/css">
                 tr.disablehover:hover {
                 background-color: white;
@@ -52,7 +52,7 @@
                                 </j:forEach>
                             </td>
                             <td>
-                                <input type="button" value="stop" onclick="stop('${res.id}')"></input>
+                                <input type="button" class="stop-container" value="stop" data-id="${res.id}"></input>
                             </td>
                         </tr>
                     </j:forEach>
