--- src/main/webapp/js/docker-manage.js
@@ -1,10 +1,0 @@
-
-function stop(theId) {
-    var input = document.getElementById('stopId');
-    if (input != null) {
-        input.value = theId;
-    }
-
-    var form = document.getElementById('control');
-    form.submit();
-}
